V0.2
Here's the translation of the update with the additional information:

Added the following features:

New models.
A basic UI interface.
Start and pause menu.
Some bugs are yet to be fixed:

Continuous use of the spacebar for jumping causes animation errors.
The Esc key cannot bring up the pause menu.