V0.3
Added the following features:

New models.
UI interface images.
Option menu.
Import of sound effect files with partial references.
Some bugs are yet to be fixed:

Continuous use of the spacebar for jumping causes animation errors.
The Esc key cannot exit the pause menu.
There's an issue with sound adjustment in the Option menu.
Problem with detecting if the player is on the ground.